Able to use event filters in subscriptions
